Starting Dev branch for v0.9.6
This branch will split from the bookx master. It's intended to improve some features that I use or need to have in my site.
Not sure if Bookx is going to get much updates or not, so this should be considered an unofficial branch. Maybe it will merge, maybe not.
About the version, 0.9.6... it's a unofficial version.