Feature for multiple filters
Add the ability to stop the propagation of irrelevant url and results if the query returns only one result.
Example: An Author is selected, if there is only one publisher it wont allow filtering for the publisher to display the same result.
This could be extended to sidebox template, to show or not the select form.
Also allows to cache db resuts in active filters == 1.