Update: Add bookx families. This feature is intended to sync software management and books online, facilitating discounts and stocks sets. A family could have several publishers, a discount, a flag to have constant stock online or not even if in your warehouse is not available at the moment.
For example: Publisher A has some recent books that you can't apply more than 10% off, but you've bought or detained  a "old" stock from Publisher A that you can apply more discount. They are yours.

Families are for this situations, when mass importing, or adding in admin panel.
Note: Still on tests